epubToAudioBook/requirements.txt
beautifulsoup4==4.13.4
EbookLib==0.19
fastapi==0.110.0
kokoro==0.9.2
mutagen==1.47.0
openai==1.85.0
requests==2.32.4
edge-tts==7.0.2
pydub==0.25.1
wyoming==1.6.0
soundfile==0.13.1
python-multipart==0.0.9
torchaudio==0.13.0
uvicorn[standard]==0.29.0
aeneas==1.7.3.0
gradio==5.33.1
docker==7.1.0
sentencex==0.6.1
gradio_log==0.0.8
faster-whisper==1.0.3
whisperx>=3.1.0,<3.2
